Rename platform_device_loader's module wasm32 -> web
